Continued the refactor by creating a directory for the modules, which extend torch.nn.Module
