+++ dna/models/models.py
@@ -8,137 +8,19 @@ import pandas as pd
 from sklearn import linear_model
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 from tqdm import tqdm
 
 from dna.data import Dataset, GroupDataLoader, PMFDataset, RNNDataLoader, group_json_objects
 from dna.kND import KNearestDatasets
 from dna import utils
-
-
-F_ACTIVATIONS = {'relu': F.relu, 'leaky_relu': F.leaky_relu, 'sigmoid': F.sigmoid, 'tanh': F.tanh}
-ACTIVATIONS = {'relu': nn.ReLU, 'leaky_relu': nn.LeakyReLU, 'sigmoid': nn.Sigmoid, 'tanh': nn.Tanh}
+from dna.models.modules.dna_module import DNAModule
+from dna.models.modules.dag_lstm_mlp import DAGLSTMMLP
+from dna.models.modules.hidden_mlp_dag_lstm_mlp import HiddenMLPDAGLSTMMLP
+from dna.models.modules.pmf import PMF
 
 
 class ModelNotFitError(Exception):
     pass
-
-
-class PyTorchRandomStateContext:
-
-    def __init__(self, seed):
-        self.seed = seed
-        self._state = None
-
-    def __enter__(self):
-        self._state = torch.random.get_rng_state()
-        torch.manual_seed(self.seed)
-
-    def __exit__(self, *args):
-        torch.random.set_rng_state(self._state)
-
-
-class Submodule(nn.Module):
-
-    def __init__(
-        self, layer_sizes: typing.List[int], activation_name: str, use_batch_norm: bool, use_skip: bool = False,
-        dropout: float = 0.0, *, device: str = 'cuda:0', seed: int = 0
-    ):
-        super().__init__()
-
-        with PyTorchRandomStateContext(seed):
-            n_layers = len(layer_sizes) - 1
-            activation = ACTIVATIONS[activation_name]
-
-            layers = []
-            for i in range(n_layers):
-                if i > 0:
-                    layers.append(activation())
-                    if dropout > 0.0:
-                        layers.append(nn.Dropout(p=dropout))
-                if use_batch_norm:
-                    layers.append(nn.BatchNorm1d(layer_sizes[i]))
-                layers.append(nn.Linear(layer_sizes[i], layer_sizes[i+1]))
-
-            self.net = nn.Sequential(*layers)
-            self.net.to(device=device)
-
-            if use_skip:
-                if layer_sizes[0] == layer_sizes[-1]:
-                    self.skip = nn.Sequential()
-                else:
-                    self.skip = nn.Linear(layer_sizes[0], layer_sizes[-1])
-                self.skip.to(device=device)
-            else:
-                self.skip = None
-
-    def forward(self, x):
-        if self.skip is None:
-            return self.net(x)
-        else:
-            return self.net(x) + self.skip(x)
-
-
-class DNAModule(nn.Module):
-
-    def __init__(
-        self, submodule_input_sizes: typing.Dict[str, int], n_layers: int, input_layer_size: int, hidden_layer_size: int,
-        output_layer_size: int, activation_name: str, use_batch_norm: bool, use_skip: bool = False, dropout: float = 0.0,
-        *, device: str = 'cuda:0', seed: int = 0
-    ):
-        super().__init__()
-        self.submodule_input_sizes = submodule_input_sizes
-        self.n_layers = n_layers
-        self.input_layer_size = input_layer_size
-        self.hidden_layer_size = hidden_layer_size
-        self.output_layer_size = output_layer_size
-        self.activation_name = activation_name
-        self._activation = F_ACTIVATIONS[activation_name]
-        self.use_batch_norm = use_batch_norm
-        self.use_skip = use_skip
-        self.dropout = dropout
-        self.device = device
-        self.seed = seed
-        self._input_seed = seed + 1
-        self._output_seed = seed + 2
-        self._dna_base_seed = seed + 3
-        self._input_submodule = self._get_input_submodule()
-        self._output_submodule = self._get_output_submodule()
-        self._dynamic_submodules = self._get_dynamic_submodules()
-
-    def _get_input_submodule(self):
-        layer_sizes = [self.input_layer_size] + [self.hidden_layer_size] * (self.n_layers - 1)
-        return Submodule(
-            layer_sizes, self.activation_name, self.use_batch_norm, self.use_skip, self.dropout, device=self.device,
-            seed=self._input_seed
-        )
-
-    def _get_output_submodule(self):
-        layer_sizes = [self.hidden_layer_size] * (self.n_layers - 1) + [self.output_layer_size]
-        return Submodule(
-            layer_sizes, self.activation_name, self.use_batch_norm, self.use_skip, self.dropout, device=self.device,
-            seed=self._output_seed
-        )
-
-    def _get_dynamic_submodules(self):
-        dynamic_submodules = torch.nn.ModuleDict()
-        for i, (submodule_id, submodule_input_size) in enumerate(sorted(self.submodule_input_sizes.items())):
-            layer_sizes = [self.hidden_layer_size * submodule_input_size] + [self.hidden_layer_size] * (self.n_layers - 1)
-            dynamic_submodules[submodule_id] = Submodule(
-                layer_sizes, self.activation_name, self.use_batch_norm, self.use_skip, self.dropout, device=self.device,
-                seed=self._dna_base_seed + i
-            )
-        return dynamic_submodules
-
-    def forward(self, args):
-        pipeline_id, pipeline, x = args
-        outputs = {'inputs.0': self._input_submodule(x)}
-        for i, step in enumerate(pipeline['steps']):
-            inputs = torch.cat(tuple(outputs[j] for j in step['inputs']), dim=1)
-            submodule = self._dynamic_submodules[step['name']]
-            h = self._activation(submodule(inputs))
-            outputs[i] = h
-        return torch.squeeze(self._output_submodule(h))
 
 
 class ModelBase:
@@ -415,185 +297,6 @@ class DNARegressionModel(PyTorchRegressionRankSubsetModelBase):
             shuffle = shuffle,
             seed = self.seed + 2
         )
-
-
-class DAGLSTM(nn.Module):
-    """
-    A modified LSTM that handles directed acyclic graphs (DAGs) by reusing and aggregating hidden states.
-    """
-
-    def __init__(
-        self, input_size: int, hidden_size: int, n_layers: int, dropout: float, reduction: str, *, device: str,
-        seed: int
-    ):
-        super().__init__()
-
-        if reduction == 'mean':
-            self.reduction = torch.mean
-        elif reduction == 'sum':
-            self.reduction = torch.sum
-        else:
-            raise Exception('No valid hidden state reduction was provided to the DAG LSTM\n'
-                            'Got \"' + reduction + '\"')
-
-        if dropout > 0:
-            # Disable cuDNN so that the LSTM layer is deterministic, see https://github.com/pytorch/pytorch/issues/18110
-            torch.backends.cudnn.enabled = False
-
-        with PyTorchRandomStateContext(seed=seed):
-            # TODO: use LSTMCell
-            self.lstm = nn.LSTM(
-                input_size=input_size, hidden_size=hidden_size, num_layers=n_layers, batch_first=True, dropout=dropout
-            )
-        self.lstm.to(device=device)
-
-    def forward(self, dags, dag_structure, lstm_start_state):
-        assert len(dag_structure) > 0
-        assert len(dag_structure) == dags.shape[1]
-
-        prev_lstm_states = {'inputs.0': lstm_start_state}  # TODO: use something better than 'inputs.0'
-        for i, node_inputs in enumerate(dag_structure):
-            # TODO: use LSTMCell
-            # uses lstm with sequence of len 1
-            nodes_i = dags[:, i, :].unsqueeze(dim=1)  # sequence of len 1
-            lstm_input_state = self._get_lstm_state(prev_lstm_states, node_inputs)
-            lstm_output, lstm_output_state = self.lstm(nodes_i, lstm_input_state)
-            prev_lstm_states[i] = lstm_output_state
-
-        return lstm_output.squeeze(dim=1)
-
-    def _get_lstm_state(self, prev_lstm_states, node_inputs):
-        """
-        Computes the aggregate hidden state for a node in the DAG.
-        """
-        mean_hidden_state = self.reduction(
-            torch.stack(
-                [prev_lstm_states[i][0] for i in node_inputs]
-            ), dim=0
-        )
-        mean_cell_state = self.reduction(
-            torch.stack(
-                [prev_lstm_states[i][1] for i in node_inputs]
-            ), dim=0
-        )
-        return (mean_hidden_state, mean_cell_state)
-
-
-class DAGLSTMMLP(nn.Module):
-    """
-    A DAG LSTM MLP embeds a DAG, concatenates that embedding with another feature vector, and passes the concatenated
-    features to an MLP.
-    """
-
-    def __init__(
-        self, lstm_input_size: int, lstm_hidden_state_size: int, lstm_n_layers: int, dropout: float,
-        mlp_extra_input_size: int, mlp_hidden_layer_size: int, mlp_n_hidden_layers: int, mlp_activation_name: str,
-        output_size: int, mlp_use_batch_norm: bool, mlp_use_skip: bool, reduction: str, *, device: str, seed: int,
-    ):
-        super().__init__()
-
-        self.device = device
-        self.seed = seed
-        self._lstm_seed = seed + 1
-        self._mlp_seed = seed + 2
-
-        self._dag_lstm = DAGLSTM(
-            lstm_input_size, lstm_hidden_state_size, lstm_n_layers, dropout, reduction, device=self.device,
-            seed=self._lstm_seed
-        )
-        self.lstm_hidden_state_size = lstm_hidden_state_size
-        self.lstm_n_layers = lstm_n_layers
-
-        mlp_input_size = lstm_hidden_state_size + mlp_extra_input_size
-        mlp_layer_sizes = [mlp_input_size] + [mlp_hidden_layer_size] * mlp_n_hidden_layers + [output_size]
-        self._mlp = Submodule(
-            mlp_layer_sizes, mlp_activation_name, mlp_use_batch_norm, mlp_use_skip, dropout, device=self.device,
-            seed=self._mlp_seed,
-        )
-
-    def forward(self, args):
-        (dag_structure, dags, features) = args
-
-        batch_size = dags.shape[0]
-        assert len(features) == batch_size, 'DAG batch size does not match features batch size'
-
-        hidden_state = torch.zeros(
-            self.lstm_n_layers, batch_size, self.lstm_hidden_state_size, device=self.device
-        )
-
-        dag_rnn_output = self._dag_lstm(dags, dag_structure, (hidden_state, hidden_state))
-        fc_input = torch.cat((dag_rnn_output, features), dim=1)
-
-        return self._mlp(fc_input).squeeze()
-
-
-class HiddenMLPDAGLSTMMLP(nn.Module):
-    """
-    The HiddenMLPDAGLSTMMLP combines a feature vector and a DAG using an MLP and a DAGLSTM and passes the final
-    embedding to an output MLP. The feature vector is first transformed using an input MLP and is then used as the
-    initial hidden state of the DAGLSTM. The DAGLSTM then creates the final embedding.
-    """
-
-    def __init__(
-        self, lstm_input_size: int, lstm_hidden_state_size: int, lstm_n_layers: int, dropout: float,
-        input_mlp_input_size: int, mlp_hidden_layer_size: int, mlp_n_hidden_layers: int, mlp_activation_name: str,
-        output_size: int, mlp_use_batch_norm: bool, mlp_use_skip: bool, reduction: str, *, device: str, seed: int,
-    ):
-        super().__init__()
-
-        self.device = device
-        self.seed = seed
-        self._input_mlp_seed = seed + 1
-        self._lstm_seed = seed + 2
-        self._output_mlp_seed = seed + 3
-
-        input_mlp_layer_sizes = [input_mlp_input_size] + [mlp_hidden_layer_size] * mlp_n_hidden_layers + [lstm_hidden_state_size]
-        input_layers = [
-            Submodule(
-                input_mlp_layer_sizes, mlp_activation_name, mlp_use_batch_norm, mlp_use_skip, dropout,
-                device=self.device, seed=self._input_mlp_seed
-            ),
-            ACTIVATIONS[mlp_activation_name]()
-        ]
-        if dropout > 0.0:
-            input_dropout = nn.Dropout(p=dropout)
-            input_dropout.to(device=device)
-            input_layers.append(input_dropout)
-        if mlp_use_batch_norm:
-            with PyTorchRandomStateContext(seed=seed):
-                input_batch_norm = nn.BatchNorm1d(lstm_hidden_state_size)
-                input_batch_norm.to(device=device)
-                input_layers.append(input_batch_norm)
-        self._input_mlp = nn.Sequential(*input_layers)
-
-        self._dag_lstm = DAGLSTM(
-            lstm_input_size, lstm_hidden_state_size, lstm_n_layers, dropout, reduction, device=self.device,
-            seed=self._lstm_seed
-        )
-        self.lstm_hidden_state_size = lstm_hidden_state_size
-        self.lstm_n_layers = lstm_n_layers
-
-        output_mlp_input_size = lstm_hidden_state_size
-        output_mlp_layer_sizes = [output_mlp_input_size] + [mlp_hidden_layer_size] * mlp_n_hidden_layers + [output_size]
-        self._output_mlp = Submodule(
-            output_mlp_layer_sizes, mlp_activation_name, mlp_use_batch_norm, mlp_use_skip, dropout, device=self.device,
-            seed=self._output_mlp_seed,
-        )
-
-    def forward(self, args):
-        (dag_structure, dags, features) = args
-
-        batch_size = dags.shape[0]
-        assert len(features) == batch_size
-
-        lstm_start_state = self.init_hidden_and_cell_state(features)
-        fc_input = self._dag_lstm(dags, dag_structure, lstm_start_state)
-        return self._output_mlp(fc_input).squeeze()
-
-    def init_hidden_and_cell_state(self, features):
-        single_hidden_state = self._input_mlp(features)
-        hidden_state = single_hidden_state.unsqueeze(dim=0).expand(self.lstm_n_layers, *single_hidden_state.shape)
-        return (hidden_state, hidden_state)
 
 
 class DAGLSTMRegressionModel(PyTorchRegressionRankSubsetModelBase):
@@ -1240,61 +943,6 @@ class ProbabilisticMatrixFactorization(PyTorchRegressionRankSubsetModelBase):
         pipeline_vec[self.pipeline_id_mapper[pipeline_id]] = 1
         pipeline_vec = torch.tensor(pipeline_vec.astype("int64"), device=self.device).long()
         return pipeline_vec
-
-
-class PMF(nn.Module):
-    """
-    In usual Matrix Factorization terms:
-    dataset_features (aka dataset_id) = movie_features
-    pipeline_features (aka pipeline_id) = user_features
-    Imagine it as a matrix like so:
-
-             Datasets (id)
-         ______________________
-    P   | ? .2  ?  ?  .4  ?  ? |
-    i   | ?  ?  ?  ?   ? .4  ? |
-    p   |.7  ?  ?  ?  .8  ?  ? |
-    e   | ? .1  ?  ?   ?  ?  ? |
-    l   | ?  ? .9  ?   ?  0  ? |
-    i   | ?  ?  ?  ?   ?  ?  ? |
-    n   |.3  ?  ?  ?  .2  ?  1 |
-    e   |______________________|
-
-
-    """
-    def __init__(self, n_pipelines, n_datasets, n_factors, device, seed):
-        super(PMF, self).__init__()
-        self.n_pipelines = n_pipelines
-        self.n_datasets = n_datasets
-        self.n_factors = n_factors
-        self.device = device
-        assert type(n_pipelines) == int and type(n_datasets) == int and type(n_factors) == int, "given wrong input for PMF: expected int"
-
-        with PyTorchRandomStateContext(seed):
-            self.pipeline_factors = torch.nn.Embedding(n_pipelines,
-                                                n_factors,
-                                                sparse=False).to(self.device)
-
-            self.dataset_factors = torch.nn.Embedding(n_datasets,
-                                                n_factors,
-                                                sparse=False).to(self.device)
-
-    def forward(self, args):
-        pipeline_id, pipeline, x = args
-        # gather embedding vectors
-        pipeline_vec = pipeline["pipeline_embedding"].to(self.device)
-        dataset_vec = x.long().to(self.device)
-        # Find the embedding id
-        dataset_nums = dataset_vec.argmax(1).tolist()
-        pipeline_num = pipeline_vec.argmax(0).tolist()
-        # find the matrix
-        matrices = torch.matmul(self.pipeline_factors(pipeline_vec), self.dataset_factors(dataset_vec).permute([0, 2, 1])).to(self.device).float()
-        # find the predicted values for each pipeline x dataset in our batch
-        subset = [tuple((index, pipeline_num, dataset_nums[index])) for index in range(matrices.shape[0])]
-        vecs = [matrices[sub].reshape(1) for sub in subset]
-        # combined the predictions
-        combined = torch.cat(vecs)
-        return combined
 
 
 def get_model(model_name: str, model_config: typing.Dict, seed: int):
